Consolidate streamed extraction into utility module with validation
Move extract_features_streamed, _maybe_rescale_bitmap, PerImageExtractionResult,
and compare_databases into pysfm/streamed_pipeline.py as a leaf utility module.
Add validate field to RigReconConfig that runs both black-box and streamed
extraction on separate databases and asserts equivalence. Remove callback
parameter (Rerun logging will be added directly in the loop). Delete the
duplicate pycolmap_recon_streamed.py pipeline.
